Platform/RPi5: Add PCIe support
This introduces PCIe support to both UEFI and ACPI.

Moreover, it also fixes RP1 USB corruption issues, caused by the
bridge's non-prefetchable memory aperture essentially creating a hole in
the PCIe's view of system RAM (mapped 1:1), through which inbound (DMA)
traffic could not pass anymore.

Therefore, we now reserve up to 1 GB of RAM for 32-bit BARs, which is
automatically reclaimed depending on how much space the connected
devices actually consume. A minimum of 5 MBs are reserved by RP1.
This is only done when ACPI is enabled; FDT uses DMA translation and can
map the entire RAM.

ECAM is still a small 4 KB window, so only a single device function can
be exposed to Windows and Linux. This is possible thanks to a few OS
workarounds:
- for Windows: "NXPMX6" OEM ID in FADT, bus 0 in MCFG is limited to
devfn 0.

- for Linux: "AMAZON" OEM ID + "GRAVITON" OEM Table ID in MCFG. An
"AMZN0001" device with _UID matching the RC segment number returns the
ECAM window in _CRS.

TF-A implements the DEN0115 spec, enabling FreeBSD, NetBSD and ESXi to
scan the entire config space.

The compatibility mode used can be configured in the setup menu. Default
is "Auto (NXPMX6 / Arm DEN0115)", which selects the NXPMX6 mode when
Windows is booted and DEN0115 for other OSes.

Tested with NVME and SATA boot + some Wi-Fi cards.

Signed-off-by: Mario Bălănică <mariobalanica02@gmail.com>